♻️ rename startActionCollection to trackActions
