+++ packages/rum/src/boot/rum.spec.ts
@@ -4,7 +4,7 @@ import { setup, TestSetupBuilder } from '../../test/specHelper'
 import { RumPerformanceNavigationTiming } from '../browser/performanceCollection'
 
 import { LifeCycleEventType } from '../domain/lifeCycle'
-import { AutoAction } from '../domain/rumEventsCollection/action/actionCollection'
+import { AutoAction } from '../domain/rumEventsCollection/action/trackActions'
 import { SESSION_KEEP_ALIVE_INTERVAL, THROTTLE_VIEW_UPDATE_PERIOD } from '../domain/rumEventsCollection/view/trackViews'
 import { RumEvent } from '../index'
 import { doGetInternalContext } from './rum'
