+++ packages/rum/src/domain/parentContexts.ts
@@ -2,7 +2,7 @@ import { monitor, ONE_MINUTE, SESSION_TIME_OUT_DELAY } from '@datadog/browser-co
 import { ActionContext, ViewContext } from '../types'
 import { ActionContextV2, ViewContextV2 } from '../typesV2'
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
-import { AutoActionCreatedEvent, AutoAction } from './rumEventsCollection/action/actionCollection'
+import { AutoAction, AutoActionCreatedEvent } from './rumEventsCollection/action/trackActions'
 import { ViewCreatedEvent } from './rumEventsCollection/view/trackViews'
 import { RumSession } from './rumSession'
