+++ packages/rum/src/types.ts
@@ -1,5 +1,5 @@
 import { Context, ErrorContext, HttpContext, ResourceType } from '@datadog/browser-core'
-import { ActionType, ActionCounts } from './domain/rumEventsCollection/action/actionCollection'
+import { ActionCounts, ActionType } from './domain/rumEventsCollection/action/trackActions'
 import { PerformanceResourceDetails } from './domain/rumEventsCollection/resource/resourceUtils'
 import { Timings, ViewLoadingType } from './domain/rumEventsCollection/view/trackViews'
 import { EventCounts } from './domain/trackEventCounts'
